feat: refactoring import getStreak
유저 스탯 관리 함수들 util로 이동
